Ensure group list doesn't become unusable when long
This commit ensures that the groups list is scrollable when it is forced
to be smaller than the viewport due to the number of groups in it.

It's not an ideal solution, as in some contexts there isn't much of a
hint to the user that the widget is scrollable. It probably beats the
existing behaviour, though.

Fixes #2971.