Add a group membership validation check at create time
This is currently enforced by an ACL on the `Annotations` resource in
`h.api`, but in view of some refactoring I have planned, I want to move
the enforcement of a check that the user can only create annotations in
groups they are a member of into the validation layer.

When the refactoring is complete, this will result in an improved error
message if a user tries to create an annotation in a group they're not a
member of via the API -- they'll get an explicit explanation of what
they have done wrong rather than a 404.