Use a mixin rather than @extend for style composition
Using @extend with a class declared in a different source file
prevents local reasoning about the final set of rules
that will apply for a class because that depends on
what order the components are included.

See http://csswizardry.com/2014/11/when-to-use-extend-when-to-use-a-mixin/
for additional reasons to avoid @extend in most cases.