Try to find annotation before importing as a new one
We sort annotations that we want to import in descending `updated` order.
Which means that when we import annotation X on the first run, and this
annotation gets updated before the second run, we will have to try and
find annotations, and on success update annotation X instead of creating
a new one which would lead to a unique constraint error on the `id`
column.

This use case needs to be handled as we want to run the migration
multiple times to ensure we have all the new and updated annotations
correctly migrated to Postgres.