Don't use pyramid_tm for the websocket server
The websocket server doesn't handle clients in a request/response cycle,
so using pyramid_tm doesn't make a lot of sense. This commit makes it
possible to use `h.db` without pyramid_tm, by deferring the
configuration of zope.sqlalchemy hooks until the session is created, and
the only if `request.tm` is present.