Fix uncaught promise rejection warning in annotation test
The change in e5bf6d1e to use the same Promise polyfill
in the tests as in the app exposed a test which tried
to call a missing stub (flash.error).

The error previously went unnoticed because it was reported
via an unhandled Promise rejection. The core-js Promise polyfill
however has a last-resort handler for uncaught promises so
the error showed up.