Improve concurrency control in websocket streamer
This is a rewrite of the core message handling routines of the streamer
to address a fundamental problem with the previous implementation.

Prior to this commit, WebSocket clients shared a thread-local
transaction manager provided by `pyramid_tm`. Access to the database was
wrapped by this transaction manager, and under high load it was possible
for incoming messages from two different websocket clients (A and B) to
trigger the following series of events:

     1. Message from A arrives, streamer starts a database transaction.
     2. Message from B arrives.
     3. Gevent context switches from the greenlet associated with client
        A to the greenlet associated with client B.
     4. Message from B queries the database, and reuses the transaction
        originally opened for client A.
     5. Processing for B's message completes, and the transaction
        commits.
     6. Gevent context switches back to the greenlet associated with
        client A.
     7. Processing for A's message completes, and the transaction
        attempts to commit.
     8. Code inside `pyramid_tm` explodes because the transaction in
        question no longer exists.

This commit ensures that all incoming messages (whether from WebSocket
clients or subscribed NSQ topics) are pushed into a
`gevent.queue.Queue`, with a single greenlet responsible for pulling
messages out of the queue. This single greenlet is also responsible for
ensuring that the database session is appropriately commited/rolled back
and closed after each message has been handled.

Since all database access from the streamer is intended to be read-only,
this is enforced with an appropriate SET TRANSACTION call before
starting message processing.

Previously, even relatively light traffic (10 connections opened at
once) was enough to precipitate the race condition described above. It
is now possible to open (at least) many hundreds of websocket
connections to the server with no problems.