Switch from python-statsd to statsd
Unfortunately, python-statsd's implementation is broken -- it calls
`socket.connect` unnecessarily, which means that on the loopback
interface (i.e. in development) stats sent to a port on which nobody is
listening will result in ICMP "destination unreachable" packets to be
returned and processed by the Python process. This in turn means that
subsequent stats submissions will log exceptions.

Use statsd instead, which does the right thing: silently sends datagrams
without regard for whether anyone is listening.