Immediately remove closed websocket clients from weakset
When a client closes the connection, immediately remove the WebSocket
instance from the instances list, rather than waiting for the heartbeat
to time the connection out and garbage collect it.

This is a small optimisation to ensure that the stats on number of
connected clients are as accurate as possible, and that we don't waste
CPU time iterating over sockets that are terminated.