Don't give each request its own transaction manager
This is a partial revert of f0c96f2, 335048d, dd5dd94. The important
parts of these commits are retained -- it is now possible to get a
database session without having to use a transaction manager, and this
is what the websocket server still does.

Unfortunately, the change made in f0c96f2 to give each request its own
transaction manager (rather than using the thread-local default
instance) doesn't play at all well with the use of
`scoped_session().query_property()` -- specifically, it seems to be
trickier than expected to ensure that the current session is always
correctly registered with the transaction manager. I'm not currently
clear on exactly what's going on here, but suffice to say that mixing a
thread-local scoped session object with per-request transaction managers
doesn't appear to work: a committed transaction doesn't flush and commit
the underlying session.

Until we have time to reevaluate the use of the query property, revert
to a `scoped_session` session factory which is explicitly registered
with a thread-local transaction manager.