Run Python tests using tox
Run the Python test suite in an isolated virtualenv, against the
installed package.

Using tox to run tests is best practice within the Python community, as
it ensures that the tests are run against the code that will be
installed by a distribution package rather than the code in the current
directory. This helps defend against manifest omissions and other common
errors.

This will also make it substantially easier to run the tests against
multiple versions of Python in the future.

One breaking change here is precipitated by the `--pyargs` option to
py.test. This causes py.test to interpret its arguments as Python module
paths rather than filesystem paths and is necessary to ensure isolation
between the tests and the checked-out repository files. In practice this
just means that if you've been used to running

    py.test h/foo/bar/

or similar, you should now run

    tox h.foo.bar

instead.