Use Node v4.3 on Travis
Use Node v4 for consistency with what we use to
build the production assets in the Docker build
and also to enable some ES2015 goodness.

Travis does not support multiple languages
in the 'language' field, so non-default Node
versions need to be installed with nvm as
part of the install process.

Rebuilding node-sass was needed to get a version
of the native binding that matched the Node version,
even when running with a clean cache. I'm not clear
at the moment on why this was necessary.