Use Element.scrollIntoView() to scroll new annotations into view
The previous method of scrolling a newly created annotation
into view in the sidebar resulted in the input field for
the new annotation losing focus.

Note that only top level annotations have an 'id'
attribute set on the corresponding HTML element.
It might make sense to change this in future although
it would add an extra binding. For now I've just noted this.