Assume `gulp build` has been run before `hypothesis-buildext`
This changes the `buildext` command so that it doesn't shell out to gulp
itself. Instead, it assumes that prebuilt assets are available in build,
and prints an error suggesting the user run `gulp build` if it can't
find a file it expects to exist.

This makes it possible to run `gulp build` just once, and then build
multiple extensions from the built assets.