Storage code for creating annotations in Postgres
Add storage-level code for creating new annotations in Postges.

The old storage.create_annotation() that creates annotations in
Elasticsearch is moved to storage.legacy_create_annotation().
storage.create_annotation() now creates annotations in Postgres.

The views code will call legacy_create_annotation() and
create_annotation() as necessary to create annotations in Elasticsearch
or Postgres depending on the postgres feature flag.

This commit is only for the create new annotation API, not for updating
or deleting annotations in Postgres.

Building on the work already done in h.api.models to enable creating or
updating DocumentMeta and DocumentURI objects in Postgres, and the work
already done in h.api.schemas to validate and transform the annotation
data and document metadata into a format more suitable for creating
objects in our Postgres database, this commit adds the h.api.storage
logic code to create new annotations in Postgres and to create or
update Document, DocumentMeta and DocumentURI objects when a new
annotation is created.

Some "semantic" validation of the data sent by the client, that cannot
be done in h.api.schemas because it requires reading the database, is
done in h.api.storage instead.

Otherwise storage just takes the validated and transform data from
schemas as-is and throws it into the database, but making sure to update
existing equivalent objects where they already exist instead of creating
new ones.