Introduce an AnnotationBasePresenter class
Introducing a presenter base class will make it easier to ensure that
the various parts of the application that need access to links to
annotations will be able to fetch them using a consistent interface.

This also changes the signature of the presenter constructor so that a
request is passed in, allowing URL generating and feature-flag checking
within presenters.