Switch to the group containing the selected annotation
When the sidebar loads as a result of an '#annotations'
fragment, we do not know which group that annotation is in.

Handle this by first fetching annotations in all groups,
then finding the group containing the selected annotation(s)
and finally filtering the results by that group.

An alternative approach would be to first fetch the
selected annotations in order to determine the group and
then fetching annotations in that group.

This approach however avoids an extra round-trip to the server
in the common case where the total number of annotations across
all groups is less than the default chunk size (200).