Introduce annotation "link generators"
Rather than hard-coding all the different types of link in the API
presenter layer, we want to allow reusers of `h.api` to add their own
hypermedia links to the rendered annotations, referring to services of
their own.

To enable this, we add a configuration directive,
`add_annotation_link_generator`, which allows such users (and ourselves)
to register named link generators for annotations, which are defined as
callables accepting the current request and the annotation for which to
generate a link.