Include `h.api` module first at application boot
In order that configuration directives like the newly-added
`add_annotation_link_generator` are available to other modules in the
web application, we have to include `h.api` first. This seems unlikely
to change as such dependencies should always be dependencies of the "app
code" on the `h.api` module, and never the reverse.