Split out elastic.DocumentMeta type normalisation
Due to Elasticsearch's dynamic mapping creation, we can't normalise the
type string of an elastic DocumentMeta. Creating an annotation would
still work since it doesn't run this code, but updating an existing
annotation fails when it tries do create or update the Document ES
document because it uses the normalised version which then clashes with
the mapping that Elasticsearch automatically created based on old data.

I've moved the normalisation part into its own method
(`normalized_type`) in order to still be able to use it in the
migration, and thus automatically clean up our document metadata when we
migrate it from Elasticsearch to Postgres.

This bug was introduced when we merged #3011 and used the new
presentation layer for rendering Elasticsearch annotations.

Fixes #3062