Refactor and speed up client load
embed.js created a public but undocumented window.hypothesisInstall()
function which had functionality ostensibly for customizing the
injection method. This was unused however.

What was much worse is that <script> and <style> tags were loaded
_serially_ by adding each <script> tag to the page and then waiting for
the `onload` callback to be invoked before adding the next. This added
needless latency to app startup.

This commit makes the install function private, simplifies it and adds
all script and style tags to the page immediately so that the browser
can load resources concurrently.