Use existing routes_mapper fixture to fix tests
WebOb 1.6 was just released, which includes a security measure to ensure
that `Location` headers do not contain newlines or carriage returns.
This means that locations need to be iterable objects (so they can be
used with `in`).

Rather than using a basic `Mock` for the `route_url` function, use the
existing `routes_mapper` fixture.