Enforce execution order of scripts in embed.js
The previous commit enables parallel bundle loading, however we
still need to enforce that the bundles are executed in the correct
order, otherwise the app bundle (injector.bundle.js) may execute before
the libraries it depends upon have loaded.

Script-inserted <script> tags default to being async, which means they
are executed as soon as they are loaded.

Setting this flag to false enables parallel loading but maintains
execution order as if the tags were encountered in HTML markup.