Unload all current annotations when new frames connect
The sidebar supports multiple client pages connecting to it
and displaying annotations for all of the connected clients.
As far as I am aware, this functionality is not actually used
anywhere, but this commit fixes an issue where annotations
would be loaded multiple times for some clients if multiple
clients connected.