Refactor feature flag to access through Client
Which allows us to keep the same usage
`request.feature('new-feature-name')`, but the database access is much
more optimised.

As soon as `request.feature` is accessed, we go to the database and read
the state of all feature flags into memory, Pyramid then memoizes the
client and subsequent calls to check if a given feature flag is enabled
or not are going directly to the internal cache within the request
lifecycle.