Ensure cursor is positioned after block prefix when creating lists
Improve behavior when the user clicks the toolbar buttons to create a
list or quote and the cursor is positioned at the start of the line with
no selection. In this case, after the command is applied, the cursor
should be positioned ready for the user to start typing the quote or
list.

When the selection is empty and text is inserted at the cursor position,
it is ambiguous whether the selection is 'before' or 'after' the
replaced text. The fix here is to treat the selection as being 'before'
the replaced text.

Fixes #3091