Refactor feature caching to be lazy-loaded
Instead of loading all features from the database when the Client is
initialized, this will only load the features from the database when the
actual state of the flags is accessed through either `enabled` or `all`.

The reason for this refactoring is that we have long-lived requests in
the `streamer` and package, and for the workers. These requests live for
as long as the Python process does, which means that we have to manually
invalidate the cache when a new chunk of work is about to get handled.

With the old caching mechanism, this meant that the only way to
invalidate the cache is to load the state of the flags from the
database. This solution however adds more roundtrips to Postgres, since
most websocket work and worker jobs don't rely on features at the
moment, so there's no need to load them from the database. But to
prevent weird behaviour in the future, there is a need to invalidate the
cache.

This commit changes the caching mechanism to only load the state of the
flags from the database when the state of one particular feature is
accessed. Thus allowing us to have a `clear()` method that only resets
the internal caching variable to an empty dictionary.

The methods that rely on the cache being loaded will now first check if
the cache is already loaded, if not, they will call `load()` first, and
then access the flags through the fresh cache.