Remove status bitfield helper properties from User model
These are all legacy properties, and unused. In some cases
(`subscriptions`) it's not even clear that they were ever meaningfully
used.

Each of `email_confirmed`, `optout`, and `invited` appear nowhere else
in the code. `subscriptions` is referenced (and set) by
`h.notifications` but appears to serve no purpose.

The field in the database will be removed in future commits.