Parse document metadata in schemas.py
parse_document_claims.py is a standalone module that's called by
h.api.schemas to transform the "document" sub-objects that the client
POSTs in annotation create and update requests into a format convenient
for initializing DocumentURI and DocumentMeta objects.

The functions in parse_document_claims just take dicts (the document
dicts from the client) as input and return lists of document uri and
document meta dicts as output. But those document uri and document meta
dicts can be passed directly to h.api.model's factory functions for
creating DocumentURIs and DocumentMetas with **.