Move annotation data transformation into schemas
Move most of the code for validating and transforming the annotation
data the client sends into the format required by storage, into
h.api.schemas.

This is for when creating new annotations in Postgres, not for the
legacy Elasticsearch "database". Since I don't want to change the legacy
behaviour at all, I split CreateAnnotationSchema.validate() into
separate methods for when the postgres_write feature is on and off.

Once the postgres_write feature is removed,
CreateAnnotationSchema._legacy_validate() can be deleted and
CreateAnnotationSchema._validate() can be moved to
CreateAnnotationSchema.validate().

The idea is that h.api.schemas should do any _structural_ validation
_and transformation_ of the data from the client. So things like
renaming 'uri' to 'target_uri', changing the complex 'permissions'
object into a simple 'shared' boolean, etc move into schemas.

On the other hand any "semantic" validation or transformation (which for
the purposes of this commit means any work that requires looking in the
database) should be done in h.api.storage not in schemas. For one thing
we don't want schemas to be accessing the database.

So for example setting the "groupid" of replies to be the same as that
of their parent annotations is done in storage, because it requires
looking up the parent annotation in the database.

Also validating that the request has permission to create an annotation
in the group is done in storage because this has to be done _after_ the
groupid has been set for replies.