Add more validation when postgres_write is on
Add more structural validation to the create annotation API endpoint
when the postgres_write feature flag is on.

This is intended to be just the minimal amount of validation to prevent
the code from crashing when users throw invalid data (required keys
missing, values of the wrong types) at the API.

Rather than extending the existing AnnotationSchema we move
AnnotationSchema to LegacyAnnotationSchema and create a new schema
called AnnotationSchema. The new AnnotationSchema is only used by the
create annotation endpoint and only when postgres_write is on.

Otherwise we would have to fix a lot of tests that call the create and
update annotation API endpoints with postgres_write off and pass invalid
data to them - dozens of tests would now be getting validation errors
that they weren't getting before.