Use the given db session no the class's one
Use the given session argument in create_or_update_document_uri() and
create_or_update_document_meta(), instead of the query properties of the
DocumentMeta and DocumentURI classes.

This means that we're no longer relying on the fact that the session
argument is the same db session as the one used by Document.query.
(For example in the websocket process these would be different database
sessions.)

In the process of doing this also fix a couple of tests to use the real
test database session not a mock one. It's easier to use the real db
than to mock sqlalchemy's complex API, I don't want to have to do things
like:

    mock_db_session.query.return_value.filter.return_value.one_or_none.return_value = document_meta

These tests were using both the real and a mock db session in the same
test which is weird. Now they just use the real test db.