Transform annotation data in schemas.py
Add a new CreateAnnotationSchema that does much of the validation and
transformation of the data that the clients POSTs in annotation create
requests.

Schemas does all of the validation and transformation that it can do to
ensure that the data is valid and in a convenient format for storing in
the database, without accessing the database.

Any validation or transformation that does require accessing the
database will have to be done later, by the storage module, because the
schemas module doesn't have access to the database. For example, setting
the groupid of replies to be the same as that of their parent
annotation. Or checking that a reply's parent annotation ID is actual
the ID of an annotation that exists in our database.

This pull request only validates and transforms the annotation data from
the client, it does not yet handle the document metadata that the client
sends in the "document" sub-dict.

We split out LegacyCreateAnnotationSchema which maintains the old
behaviour the CreateAnnotationSchema used to have. The views code calls
this LegacyCreateAnnotationSchema instead of the new
CreateAnnotationSchema. The new CreateAnnotationSchema is not in use yet
as more code needs to be put in place (e.g. in storage) before it can be
used.