Add more validation of create annotation requests
Add a lot more validation of create annotation requests.

This is intended to be just the minimal amount of validation to prevent
the code from crashing now that we're saving the annotations in
PostgreSQL instead of in Elasticsearch.

The new validation only happens for create annotation requests (not
update) and only when the postgres feature flag is on.

The new validation is added to AnnotationSchema.
LegacyAnnotationSchema is split out from AnnotationSchema and has the
same behaviour as AnnotationSchema used to have (i.e. it doesn't have
the new additional validation).

LegacyCreateAnnotationSchema (which is the schema that views.py uses
when the postgres feature flag is off) not calls LegacyAnnotationSchema
instead of AnnotationSchema, so the new validation isn't used when the
postgres feature flag is off.

UpdateAnnotationSchema now calls LegacyAnnotationSchema instead of
AnnotationSchema as well, so that the new validation isn't applied to
annotation updates.