Shuffle h.auth.util.bearer_token into h.auth.tokens
This is a small refactoring that moves the logic that was split between
the authentication policy and the `h.auth.util.bearer_token` function
into a single function which takes a request object and returns any
valid token-authenticated userid for that request.

This simplifies testing the authentication policy, but it's also a
cleaner interface as we no longer have to deal with null values, empty
strings, and so on in multiple places.