Use fixtures in all search tests
Use the search_lib and AnnotationJSONPresenter fixtures in all
h.api.search() tests by putting the fixture on the test class.

This ensures that any test methods (existing or new) that don't take
these fixtures as test method arguments, because they don't need to
access them, will still use the fixtures and therefore the mocked
search_lib and AnnotationJSONPresenter not the real ones.