Tighten constraints on annotation IDs
Annotation IDs are base64url-encoded UUIDs (or weird Elasticsearch flake
IDs, but that's another story). This means they are 20 or 22 ASCII
characters long and composed of the character set [A-Za-z0-9_-].

In preparation for adding another route with a file extension, this
commit tightens up the regex match for annotation IDs for the

  /api/annotations/<id>

route.