Simplify group permissions check
Simplify the check in h.api.storage.create_annnotation() that a user
must be a member of a group in order to create an annotation in that
group.

If the annotation's group is __world__ then there's nothing to check -
any authenticated user can create an annotation in the __world__ group.
User's can't call the h.api.view.create() view (which is what calls
storage.create_annnotation()) without already being authenticated, and
storage.create_annnotation() already assumes that the user is
authenticated because it assumes that data['userid'] is set (which is
set to request.authenticated_userid by h.api.schemas).