Fix crash when creating replies in postgres
When the postgres feature flag is on each new annotation is created in
both PostgreSQL and the legacy Elasticsearch database.

When the annotation being created is a reply, part of creating the
annotation involves fetching its parent annotation from the database by
calling fetch_annotation().

fetch_annotation() would fetch from either Elasticsearch or Postgres
depending on the feature flag. But when the feature flag is on
fetch_annotation() is actually called twice: once when creating the
reply in Elasticsearch and once when creating it in Postgres. Both times
fetch_annotation() would fetch the parent annotation from Postgres,
since the postgres feature flag is on. The code for creating a reply in
Elasticsearch would then crash because it's dealing with a Postgres
annotation object but it only works with Elasticsearch annotation
objects.

So add a postgres argument to fetch_annotation() to force it to fetch
the annotation from either postgres or elasticsearch.