Add a new Elasticsearch index
Add a second Elasticsearch index named "hypothesis" by default. This is
intended to be used as the new Elasticsearch index that will be filled
from annotation data in the new PostgreSQL database.

The legacy index named "annotator" by default is still present and
everything still reads from and writes to that index.

This commit adds:

- The config settings for the new index
- The code to create the index on application startup if it doesn't exist
- The code to read (i.e. search) from the new index when the 'postgres'
  feature flag is on

This commit does not add:

- Any code to write to the new index.
  The new search index is always empty for now.
  Later commits will add a worker process that writes changes to the new index.

Config settings
---------------

Rename the existing es.index configuration setting for the now legacy
Elasticsearch database to legacy.es.index.

Use the original name es.index for the new Elasticsearch index.

The new Elasticsearch index uses the same host (the es.host setting) and
other settings as the legacy one (only the index name can be set
differently for each index). This means that we don't need to make any
changes to the config settings in our deployment code - it'll just
continue to use the same Elasticsearch host and other settings, and the
legacy index's name will default to "annotator" and the new index's name
will default to "hypothesis", and the new index will simply be
automatically created in the existing Elasticsearch host when this code
is deployed.

Details:

* The es.index setting becomes legacy.es.index.
  The default value is "annotator" as before.

* A new es.index setting is added whose default value is "hypothesis".

* Other Elasticsearch settings do not get legacy copies, rather the
  single setting is used for both the legacy and the new Elasticsearch:
  es.host, es.compatibility_mode, es.client_poolsize and es.client_timeout.

request.es and request.legacy_es
--------------------------------

request.es is now an Elasticsearch client for the new search index _if_
the 'postgres' feature flag is on, but remains a client for the old
index if the feature flag is off. So searches will continue to go to the
old search index, then will go to the new search index when the feature
flag is turned on.

The new request.legacy_es is always an Elasticsearch client for the old
search index, regardless of whether the 'postgres' feature flag is on or
off. All code that _writes_ to the search index has been changed to use
this, as we always want to write to the old search index regardless of
whether the feature flag is on or off.

For example when creating, updating or deleting annotations with the
postgres feature flag on these changes will first be written to
PostgreSQL then to the legacy Elasticsearch index. If the feature flag
is off the changes will be written to the legacy Elasticsearch index
only.

Nothing writes to the new Elasticsearch index yet. Later commits will
add a worker process that writes changes to the new index.

NIPSA worker
------------

The NIPSA worker always writes to the legacy Elasticsearch index.
There's no support for NIPSA with the new index yet.

annotool
--------

The `hypothesis annotool` command always writes to the legacy
Elasticsearch index. There's no support for annotool with the new index
yet.

reindex
-------

The `hypothesis reindex` command always writes to the legacy
Elasticsearch index. There's no support for reindexing with the new index
yet.

scripts/migrate.py
------------------

The migration script has been changed to use request.legacy_es so that
it always uses the legacy Elasticsearch index, as it should.