Add a Celery application for processing background tasks
This adds a minimal Celery application which can be used to process
background tasks from the rest of the application.

Until now, we've used `h.worker`, a home-grown worker runner which reads
tasks from NSQ. It has served its purpose, but is increasingly
inflexible, as it currently assumes a process-per-job-type worker model
which makes deploying new job types tricky, and consumes more resources
as we add more job types, even if those jobs are infrequently used.

Rather than continuing to develop `h.worker`, it makes sense to switch
to the well-known and actively developed Celery project to run
background tasks. This allows us to configure tasks in a straightforward
fashion (basically just functions) and dispatch to them easily from
anywhere in the code.

Like `h.worker`, this Celery application provides integration with the
Pyramid application, by attaching a fake `request` object to the
application when a worker process starts up. This `request` object can
be used to retrieve application settings, the database session, etc.